xrp: refactor types and events
